==> Top100_Analysis_Codes/1_billboard_top100_scrape.ipynb <==
import pandas as pd
import numpy as np

# Billboard Year-End Hot 100 Charts 2000-2018

billboard100_df = pd.DataFrame(columns=['Title', 'Artist', 'Year'])
billboard100_df

url_template = 'https://en.wikipedia.org/wiki/Billboard_Year-End_Hot_100_singles_of_'

years = [2000, 2001, 2002, 2003, 2004, 2005, 2006, 
         2007, 2008, 2009, 2010, 2011, 2012, 2013,
         2014, 2015, 2016, 2017, 2018]

# 2000
url = url_template + str(2000)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2000

billboard100_df = billboard100_df.append(temp_df)

# 2001
url = url_template + str(2001)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2001

billboard100_df = billboard100_df.append(temp_df)

# 2002
url = url_template + str(2002)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2002

billboard100_df = billboard100_df.append(temp_df)

# 2003
url = url_template + str(2003)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2003

billboard100_df = billboard100_df.append(temp_df)

# 2004
url = url_template + str(2004)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2004

billboard100_df = billboard100_df.append(temp_df)


# 2005
url = url_template + str(2005)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2005

billboard100_df = billboard100_df.append(temp_df)

# 2006
url = url_template + str(2006)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2006

billboard100_df = billboard100_df.append(temp_df)

# 2007
url = url_template + str(2007)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2007

billboard100_df = billboard100_df.append(temp_df)

# 2008
url = url_template + str(2008)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2008

billboard100_df = billboard100_df.append(temp_df)

# 2009
url = url_template + str(2009)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2009

billboard100_df = billboard100_df.append(temp_df)

# 2010
url = url_template + str(2010)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2010

billboard100_df = billboard100_df.append(temp_df)

# 2011
url = url_template + str(2011)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2011

billboard100_df = billboard100_df.append(temp_df)

# 2012
url = url_template + str(2012)

temp_df = pd.read_html(url, header=0)[1]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2012

billboard100_df = billboard100_df.append(temp_df)

# 2013
url = url_template + str(2013)

temp_df = pd.read_html(url, header=0)[1]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2013

billboard100_df = billboard100_df.append(temp_df)

# 2014
url = url_template + str(2014)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2014

billboard100_df = billboard100_df.append(temp_df)

# 2015
url = url_template + str(2015)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2015

billboard100_df = billboard100_df.append(temp_df)

# 2016
url = url_template + str(2016)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2016

billboard100_df = billboard100_df.append(temp_df)

# 2017
url = url_template + str(2017)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2017

billboard100_df = billboard100_df.append(temp_df)

# 2018
url = url_template + str(2018)

temp_df = pd.read_html(url, header=0)[0]
temp_df.head()
temp_df.columns.values[2] = "Artist"
temp_df = temp_df.drop(temp_df.columns[0], axis=1)
temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
temp_df.Title = [title.strip('\"') for title in temp_df.Title]
temp_df['Year'] = 2018

billboard100_df = billboard100_df.append(temp_df)


billboard100_df = billboard100_df.reset_index(drop=True)
billboard100_df['Year'] = pd.to_numeric(billboard100_df['Year'])
billboard100_df['Title'] = billboard100_df['Title'].astype(str)
billboard100_df['Artist'] = billboard100_df['Artist'].astype(str)

billboard100_df

billboard100_df.to_csv('billboard.csv')

# # 2019
# url = url_template + str(2019)

# temp_df = pd.read_html(url, header=0)[0]

# temp_df.columns.values[2] = "Artist"

# temp_df = temp_df.drop(temp_df.columns[0], axis=1)
# temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
# temp_df.Title = [title.strip('\"') for title in temp_df.Title]
# temp_df['Year'] = 2019
# temp_df.head()
# billboard100_df = billboard100_df.append(temp_df)

# # 2020
# url = url_template + str(2020)

# temp_df = pd.read_html(url, header=0)[0]

# temp_df.columns.values[2] = "Artist"

# temp_df = temp_df.drop(temp_df.columns[0], axis=1)
# temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
# temp_df.Title = [title.strip('\"') for title in temp_df.Title]
# temp_df['Year'] = 2020
# temp_df.head()
# billboard100_df = billboard100_df.append(temp_df)

# # 2021
# url = url_template + str(2021)

# temp_df = pd.read_html(url, header=0)[0]

# temp_df.columns.values[2] = "Artist"

# temp_df = temp_df.drop(temp_df.columns[0], axis=1)
# temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
# temp_df.Title = [title.strip('\"') for title in temp_df.Title]
# temp_df['Year'] = 2021
# temp_df.head()
# billboard100_df = billboard100_df.append(temp_df)

# # 2022
# url = url_template + str(2022)

# temp_df = pd.read_html(url, header=0)[0]

# temp_df.columns.values[2] = "Artist"

# temp_df = temp_df.drop(temp_df.columns[0], axis=1)
# temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
# temp_df.Title = [title.strip('\"') for title in temp_df.Title]
# temp_df['Year'] = 2022
# temp_df.head()
# billboard100_df = billboard100_df.append(temp_df)
# # 2023
# url = url_template + str(2023)

# temp_df = pd.read_html(url, header=0)[1]

# temp_df.columns.values[2] = "Artist"

# temp_df = temp_df.drop(temp_df.columns[0], axis=1)
# temp_df.Artist = [artist.split(' featuring')[0] for artist in temp_df.Artist]
# temp_df.Title = [title.strip('\"') for title in temp_df.Title]
# temp_df['Year'] = 2023
# temp_df.head()
# billboard100_df = billboard100_df.append(temp_df)
