==> Top100_Analysis_Codes/2_combining_billboard_19000_datasets.ipynb <==
import pandas as pd

# Cleaning Spotify Songs

spotify_songs = pd.read_csv('data/19000-spotify-songs/song_info.csv')

spotify_songs = spotify_songs.drop('album_names', 1)
spotify_songs = spotify_songs.drop('playlist', 1)

spotify_songs = spotify_songs.rename(columns={"song_name": "Title", "artist_name": "Artist"})

spotify_songs = spotify_songs.drop_duplicates(subset="Title", keep="first")

# Cleaning Billboard Songs

billboard = pd.read_csv('data/billboard.csv', index_col=0)

billboard = billboard.drop('Year', 1)
billboard = billboard.drop_duplicates(subset="Title", keep="first")

# Combining DataFrames

spotify_songs['Top100'] = 0
billboard['Top100'] = 1

combined_df = pd.concat([spotify_songs, billboard], ignore_index=True, sort=True)

combined_df = combined_df.drop_duplicates(subset="Title", keep="last")

combined_df

combined_df.to_csv('data/songs.csv')
